+++ tests/test_config.py
@@ -1,18 +1,21 @@
 from click.testing import CliRunner
 
-from src.logsight_cli import cli
+from logsight_cli.logsight_cli import cli
+
+
+AUTH = ['--email', 'jorge.cardoso.pt@gmail.com', '--password', 'pijgu8-jeCjuw-segtes']
 
 
 def test_config():
     runner = CliRunner()
-    result = runner.invoke(cli, ['config'], obj={})
+    result = runner.invoke(cli, AUTH + ['config'], obj={})
     assert result.exit_code == 0
     assert 'EMAIL' in result.output
 
 
 def test_application():
     runner = CliRunner()
-    result = runner.invoke(cli, ['application', 'ls'], obj={})
+    result = runner.invoke(cli, AUTH + ['application', 'ls'], obj={})
     assert result.exit_code == 0
     assert 'APPLICATION' in result.output